requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
drf-spectacular
psycopg2-binary
celery
redis
django-celery-beat
django-filter
